requirements.txt
mcp>=1.21.2
pymssql>=2.2.7
pyodbc>=5.0.1; platform_system == "Windows"
python-dotenv>=1.2.1
